+++ SingularityService/pom.xml
@@ -616,7 +616,7 @@
                 <image>
                   <name>hubspot/singularityservice:${project.version}</name>
                   <build>
-                    <from>hubspot/singularitybase:${mesos.docker.tag}-${singularitybase.image.revision}</from>
+                    <from>openjdk:8-jre</from>
                     <maintainer>platform-infrastructure-groups@hubspot.com</maintainer>
                     <ports>
                       <port>7099</port>
