+++ SingularityService/src/main/docker/singularity.yaml
@@ -11,7 +11,7 @@ server:
         archivedLogFilenamePattern: ../logs/access-%d.log.gz
 
 mesos:
-  masterUri: http://localhost:5050/api/v1/scheduler
+  master: http://localhost:5050/api/v1/scheduler
   defaultCpus: 1
   defaultMemory: 128
   frameworkName: Singularity
