+++ SingularityService/src/main/java/com/hubspot/singularity/config/MesosConfiguration.java
@@ -8,10 +8,8 @@ import com.google.common.base.Optional;
 @JsonIgnoreProperties( ignoreUnknown = true )
 public class MesosConfiguration {
 
-  @Deprecated
-  private String master;
   @NotNull
-  private String masterUri;
+  private String master;
   @NotNull
   private String frameworkName;
   @NotNull
@@ -216,13 +214,5 @@ public class MesosConfiguration {
 
   public void setFrameworkUser(String frameworkUser) {
     this.frameworkUser = frameworkUser;
-  }
-
-  public String getMasterUri() {
-    return masterUri;
-  }
-
-  public void setMasterUri(String masterUri) {
-    this.masterUri = masterUri;
   }
 }
