+++ SingularityService/src/main/java/com/hubspot/singularity/mesos/SingularityMesosScheduler.java
@@ -1,428 +1,150 @@
 package com.hubspot.singularity.mesos;
 
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.Set;
-import java.util.Timer;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReentrantLock;
-import java.util.stream.Collectors;
 
-import javax.inject.Singleton;
-
-import org.apache.mesos.v1.Protos;
-import org.apache.mesos.v1.Protos.AgentID;
-import org.apache.mesos.v1.Protos.ExecutorID;
-import org.apache.mesos.v1.Protos.FrameworkID;
-import org.apache.mesos.v1.Protos.FrameworkInfo;
 import org.apache.mesos.v1.Protos.InverseOffer;
 import org.apache.mesos.v1.Protos.MasterInfo;
 import org.apache.mesos.v1.Protos.Offer;
 import org.apache.mesos.v1.Protos.OfferID;
-import org.apache.mesos.v1.Protos.TaskID;
 import org.apache.mesos.v1.Protos.TaskStatus;
-import org.apache.mesos.v1.scheduler.Protos.Event.Message;
-import org.apache.mesos.v1.scheduler.Protos.Event.Subscribed;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.apache.mesos.v1.scheduler.Protos;
 
-import com.codahale.metrics.annotation.Timed;
 import com.google.common.base.Optional;
-import com.google.common.base.Preconditions;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Sets;
-import com.google.inject.Inject;
-import com.google.inject.name.Named;
-import com.google.protobuf.ByteString;
-import com.hubspot.mesos.JavaUtils;
-import com.hubspot.mesos.MesosUtils;
 import com.hubspot.singularity.RequestCleanupType;
-import com.hubspot.singularity.SingularityAbort;
-import com.hubspot.singularity.SingularityAbort.AbortReason;
-import com.hubspot.singularity.SingularityAction;
-import com.hubspot.singularity.SingularityKilledTaskIdRecord;
-import com.hubspot.singularity.SingularityMainModule;
-import com.hubspot.singularity.SingularityTask;
-import com.hubspot.singularity.SingularityTaskDestroyFrameworkMessage;
 import com.hubspot.singularity.SingularityTaskId;
 import com.hubspot.singularity.TaskCleanupType;
-import com.hubspot.singularity.config.MesosConfiguration;
-import com.hubspot.singularity.config.SingularityConfiguration;
-import com.hubspot.singularity.config.UIConfiguration;
-import com.hubspot.singularity.data.DisasterManager;
-import com.hubspot.singularity.data.TaskManager;
-import com.hubspot.singularity.data.transcoders.Transcoder;
-import com.hubspot.singularity.mesos.SingularitySlaveAndRackManager.CheckResult;
-import com.hubspot.singularity.resources.UiResource;
-import com.hubspot.singularity.scheduler.SingularityLeaderCacheCoordinator;
-import com.hubspot.singularity.sentry.SingularityExceptionNotifier;
 
-@Singleton
-public class SingularityMesosScheduler extends SingularityMesosSchedulerClient {
-
-  private static final Logger LOG = LoggerFactory.getLogger(SingularityMesosScheduler.class);
-
-  private final SingularityExceptionNotifier exceptionNotifier;
-
-  private final SingularityStartup startup;
-  private final SingularityAbort abort;
-  private final SingularityLeaderCacheCoordinator leaderCacheCoordinator;
-  private final SingularityMesosFrameworkMessageHandler messageHandler;
-  private final SingularitySlaveAndRackManager slaveAndRackManager;
-  private final DisasterManager disasterManager;
-  private final OfferCache offerCache;
-  private final SingularityMesosOfferScheduler offerScheduler;
-  private final SingularityMesosStatusUpdateHandler statusUpdateHandler;
-  private final boolean offerCacheEnabled;
-  private final boolean delayWhenStatusUpdateDeltaTooLarge;
-  private final long delayWhenDeltaOverMs;
-  private final AtomicLong statusUpdateDeltaAvg;
-  private final SingularityConfiguration configuration;
-  private final MesosConfiguration mesosConfiguration;
-  private final String singularityUriBase;
-  private final TaskManager taskManager;
-  private final Transcoder<SingularityTaskDestroyFrameworkMessage> transcoder;
-
-  private final Lock stateLock;
-
-  private final SingularitySchedulerLock lock;
+public abstract class SingularityMesosScheduler {
 
   public enum SchedulerState {
     STARTUP, SUBSCRIBED, STOPPED;
   }
 
-  private volatile SchedulerState state;
-  private Timer retryTimer = null;
-  private Optional<Long> lastOfferTimestamp = Optional.absent();
-  private final AtomicBoolean connected = new AtomicBoolean(false);
-  private final AtomicReference<MasterInfo> masterInfo = new AtomicReference<>();
+  /**
+   * First event received when the scheduler subscribes. This contains the
+   * frameworkId generated by mesos that should be used if the scheduler needs
+   * to reconnect in the future.
+   *
+   * @param subscribed Data regarding your subscription
+   */
+  public abstract void subscribed(Protos.Event.Subscribed subscribed);
 
-  private final List<TaskStatus> queuedUpdates;
+  /**
+   * Received whenever there are new resources that are offered to the
+   * scheduler. Each offer corresponds to a set of resources on an agent.
+   * Until the scheduler accepts or declines an offer the resources are
+   * considered allocated to the scheduler.
+   *
+   * @param offers A list of offers from mesos
+   */
+  public abstract void resourceOffers(List<Offer> offers);
 
-  @Inject
-  SingularityMesosScheduler(SingularitySchedulerLock lock,
-                            SingularityExceptionNotifier exceptionNotifier,
-                            SingularityStartup startup,
-                            SingularityLeaderCacheCoordinator leaderCacheCoordinator,
-                            SingularityAbort abort,
-                            SingularityMesosFrameworkMessageHandler messageHandler,
-                            SingularitySlaveAndRackManager slaveAndRackManager,
-                            OfferCache offerCache,
-                            SingularityMesosOfferScheduler offerScheduler,
-                            SingularityMesosStatusUpdateHandler statusUpdateHandler,
-                            DisasterManager disasterManager,
-                            SingularityConfiguration configuration,
-                            TaskManager taskManager,
-                            Transcoder<SingularityTaskDestroyFrameworkMessage> transcoder,
-                            @Named(SingularityMainModule.STATUS_UPDATE_DELTA_30S_AVERAGE) AtomicLong statusUpdateDeltaAvg,
-                            @Named(SingularityMainModule.SINGULARITY_URI_BASE) final String singularityUriBase) {
-    this.exceptionNotifier = exceptionNotifier;
-    this.startup = startup;
-    this.abort = abort;
-    this.messageHandler = messageHandler;
-    this.slaveAndRackManager = slaveAndRackManager;
-    this.disasterManager = disasterManager;
-    this.offerCache = offerCache;
-    this.offerScheduler = offerScheduler;
-    this.statusUpdateHandler = statusUpdateHandler;
-    this.offerCacheEnabled = configuration.isCacheOffers();
-    this.delayWhenStatusUpdateDeltaTooLarge = configuration.isDelayOfferProcessingForLargeStatusUpdateDelta();
-    this.delayWhenDeltaOverMs = configuration.getDelayPollersWhenDeltaOverMs();
-    this.statusUpdateDeltaAvg = statusUpdateDeltaAvg;
-    this.taskManager = taskManager;
-    this.transcoder = transcoder;
-    this.leaderCacheCoordinator = leaderCacheCoordinator;
-    this.queuedUpdates = Lists.newArrayList();
-    this.lock = lock;
-    this.stateLock = new ReentrantLock();
-    this.state = SchedulerState.STARTUP;
-    this.singularityUriBase = singularityUriBase;
-    this.mesosConfiguration = configuration.getMesosConfiguration();
-    this.configuration = configuration;
-  }
+  /**
+   * Received whenever there are resources requested back from the scheduler.
+   * Each inverse offer specifies the agent, and optionally specific
+   * resources. Accepting or Declining an inverse offer informs the allocator
+   * of the scheduler's ability to release the specified resources without
+   * violating an SLA. If no resources are specified then all resources on the
+   * agent are requested to be released.
+   *
+   * @param offers A list of reverse offers from mesos
+   */
+  public abstract void inverseOffers(List<InverseOffer> offers);
 
-  @Override
-  public void subscribed(Subscribed subscribed) {
-    callWithLock(() -> {
-      Preconditions.checkState(state == SchedulerState.STARTUP, "Asked to startup - but in invalid state: %s", state.name());
+  /**
+   * Received when a particular offer is no longer valid (e.g., the agent
+   * corresponding to the offer has been removed) and hence needs to be
+   * rescinded. Any future calls ('Accept' / 'Decline') made by the scheduler
+   * regarding this offer will be invalid.
+   *
+   * @param offerId the recinded offer
+   */
+  public abstract void rescind(OfferID offerId);
 
-      leaderCacheCoordinator.activateLeaderCache();
-      MasterInfo newMasterInfo = subscribed.getMasterInfo();
-      masterInfo.set(newMasterInfo);
-      startup.startup(newMasterInfo);
-      stateLock.lock(); // ensure we aren't adding queued updates. calls to status updates are now blocked.
+  /**
+   * Received when a particular inverse offer is no longer valid (e.g., the
+   * agent corresponding to the offer has been removed) and hence needs to be
+   * rescinded. Any future calls ('Accept' / 'Decline') made by the scheduler
+   * regarding this inverse offer will be invalid.
+   *
+   * @param offerId The rescind inverse offer id
+   */
+  public abstract void rescindInverseOffer(OfferID offerId);
 
-      try {
-        state = SchedulerState.SUBSCRIBED;
-        queuedUpdates.forEach(statusUpdateHandler::processStatusUpdate);
-      } finally {
-        stateLock.unlock();
-      }
-    }, "subscribed", false);
-  }
+  /**
+   * Received whenever there is a status update that is generated by the
+   * executor or agent or master. Status updates should be used by executors
+   * to reliably communicate the status of the tasks that they manage. It is
+   * crucial that a terminal update (see TaskState in v1/mesos.proto) is sent
+   * by the executor as soon as the task terminates, in order for Mesos to
+   * release the resources allocated to the task. It is also the
+   * responsibility of the scheduler to explicitly acknowledge the receipt of
+   * a status update. See 'Acknowledge' in the 'Call' section below for the
+   * semantics.
+   *
+   * @param update Contains info about the current tasks status
+   */
+  public abstract void statusUpdate(TaskStatus update);
 
-  @Timed
-  @Override
-  public void resourceOffers(List<Offer> offers) {
-    if (!isRunning()) {
-      LOG.info("Scheduler is in state {}, declining {} offer(s)", state.name(), offers.size());
-      decline(offers.stream().map(Offer::getId).collect(Collectors.toList()));
-      return;
-    }
-    callWithLock(() -> {
-      final long start = System.currentTimeMillis();
-      LOG.info("Received {} offer(s)", offers.size());
-      lastOfferTimestamp = Optional.of(System.currentTimeMillis());
-      boolean delclineImmediately = false;
-      if (disasterManager.isDisabled(SingularityAction.PROCESS_OFFERS)) {
-        LOG.info("Processing offers is currently disabled, declining {} offers", offers.size());
-        delclineImmediately = true;
-      }
-      if (delayWhenStatusUpdateDeltaTooLarge && statusUpdateDeltaAvg.get() > delayWhenDeltaOverMs) {
-        LOG.info("Status update delta is too large ({}), declining offers while status updates catch up", statusUpdateDeltaAvg.get());
-        delclineImmediately = true;
-      }
+  /**
+   * Received when a custom message generated by the executor is forwarded by
+   * the master. Note that this message is not interpreted by Mesos and is
+   * only forwarded (without reliability guarantees) to the scheduler. It is
+   * up to the executor to retry if the message is dropped for any reason.
+   *
+   * @param message Message sent from executor
+   */
+  public abstract void message(Protos.Event.Message message);
 
-      if (delclineImmediately) {
-        decline(offers.stream().map(Offer::getId).collect(Collectors.toList()));
-        return;
-      }
+  /**
+   * Received when an agent is removed from the cluster (e.g., failed health
+   * checks) or when an executor is terminated. Note that, this event
+   * coincides with receipt of terminal UPDATE events for any active tasks
+   * belonging to the agent or executor and receipt of 'Rescind' events for
+   * any outstanding offers belonging to the agent. Note that there is no
+   * guaranteed order between the 'Failure', 'Update' and 'Rescind' events
+   * when an agent or executor is removed.
+   *
+   * @param failure Information regarding the current failure
+   */
+  public abstract void failure(Protos.Event.Failure failure);
 
-      if (offerCacheEnabled) {
-        if (disasterManager.isDisabled(SingularityAction.CACHE_OFFERS)) {
-          offerCache.disableOfferCache();
-        } else {
-          offerCache.enableOfferCache();
-        }
-      }
+  /**
+   * Received when there is an unrecoverable error in the scheduler (e.g.,
+   * scheduler failed over, rate limiting, authorization errors etc.). The
+   * scheduler should abort on receiving this event.
+   *
+   * @param message Error message
+   */
+  public abstract void error(String message);
 
-      List<Protos.Offer> offersToCheck = new ArrayList<>(offers);
+  /**
+   * Periodic message sent by the Mesos master according to
+   * 'Subscribed.heartbeat_interval_seconds'. If the scheduler does not
+   * receive any events (including heartbeats) for an extended period of time
+   * (e.g., 5 x heartbeat_interval_seconds), there is likely a network
+   * partition. In such a case the scheduler should close the existing
+   * subscription connection and resubscribe using a backoff strategy.
+   */
+  public abstract void heartbeat();
 
-      for (Offer offer : offers) {
-        String rolesInfo = MesosUtils.getRoles(offer).toString();
-        LOG.debug("Received offer ID {} with roles {} from {} ({}) for {} cpu(s), {} memory, {} ports, and {} disk", offer.getId().getValue(), rolesInfo, offer.getHostname(), offer.getAgentId()
-                .getValue(), MesosUtils.getNumCpus(offer), MesosUtils.getMemory(offer),
-            MesosUtils.getNumPorts(offer), MesosUtils.getDisk(offer));
+  /**
+   * Singularity specific methods used elsewhere in the code to determine scheduler
+   * state and wrap certain actions
+   */
+  public abstract SchedulerState getState();
 
-        CheckResult checkResult = slaveAndRackManager.checkOffer(offer);
-        if (checkResult == CheckResult.NOT_ACCEPTING_TASKS) {
-          decline(Collections.singletonList(offer.getId()));
-          offersToCheck.remove(offer);
-          LOG.debug("Will decline offer {}, slave {} is not currently in a state to launch tasks", offer.getId().getValue(), offer.getHostname());
-        }
-      }
+  public abstract void notifyStopping();
 
-      final Set<OfferID> acceptedOffers = Sets.newHashSetWithExpectedSize(offersToCheck.size());
+  public abstract void slaveLost(org.apache.mesos.v1.Protos.AgentID slaveId);
 
-      try {
-        List<SingularityOfferHolder> offerHolders = offerScheduler.checkOffers(offers);
+  public abstract boolean isRunning();
 
-        for (SingularityOfferHolder offerHolder : offerHolders) {
-          if (!offerHolder.getAcceptedTasks().isEmpty()) {
-            offerHolder.launchTasks(this);
+  public abstract Optional<MasterInfo> getMaster();
 
-            acceptedOffers.add(offerHolder.getOffer().getId());
-          } else {
-            offerCache.cacheOffer(start, offerHolder.getOffer());
-          }
-        }
-      } catch (Throwable t) {
-        LOG.error("Received fatal error while handling offers - will decline all available offers", t);
+  public abstract void start() throws Exception;
 
-        for (Protos.Offer offer : offersToCheck) {
-          if (acceptedOffers.contains(offer.getId())) {
-            continue;
-          }
-
-          decline(Collections.singletonList(offer.getId()));
-        }
-
-        throw t;
-      }
-
-      LOG.info("Finished handling {} new offer(s) ({}), {} accepted, {} declined/cached", offers.size(), JavaUtils.duration(start), acceptedOffers.size(),
-          offers.size() - acceptedOffers.size());
-    }, "offers");
-  }
-
-  @Override
-  public void inverseOffers(List<InverseOffer> offers) {
-    LOG.debug("Singularity is currently not able to handle inverse offers events");
-  }
-
-  @Override
-  public void rescind(OfferID offerId) {
-    callWithLock(() -> offerCache.rescindOffer(offerId), "rescind");
-  }
-
-  @Override
-  public void rescindInverseOffer(OfferID offerId) {
-    LOG.debug("Singularity is currently not able to handle inverse offers events");
-  }
-
-  @Override
-  public void statusUpdate(TaskStatus status) {
-    final long start = System.currentTimeMillis();
-    stateLock.lock();
-
-    try {
-      if (!isRunning()) {
-        LOG.info("Scheduler is in state {}, queueing an update {} - {} queued updates so far", state.name(), status, queuedUpdates.size());
-        queuedUpdates.add(status);
-        return;
-      }
-    } finally {
-      stateLock.unlock();
-    }
-
-    try {
-      statusUpdateHandler.processStatusUpdate(status);
-    } catch (Throwable t) {
-      LOG.error("Scheduler threw an uncaught exception - exiting", t);
-      exceptionNotifier.notify(String.format("Scheduler threw an uncaught exception (%s)", t.getMessage()), t);
-      abort.abort(AbortReason.UNRECOVERABLE_ERROR, Optional.of(t));
-    } finally {
-      LOG.debug("Handled status update for {} in {}", status.getTaskId().getValue(), JavaUtils.duration(start));
-    }
-  }
-
-  @Override
-  public void message(Message message) {
-    callWithLock(() -> {
-      ExecutorID executorID = message.getExecutorId();
-      AgentID slaveId = message.getAgentId();
-      byte[] data = message.getData().toByteArray();
-      LOG.info("Framework message from executor {} on slave {} with {} bytes of data", executorID, slaveId, data.length);
-      messageHandler.handleMessage(executorID, slaveId, data);
-    }, "frameworkMessage");
-  }
-
-  @Override
-  public void failure(org.apache.mesos.v1.scheduler.Protos.Event.Failure failure) {
-    if (failure.hasExecutorId()) {
-      LOG.warn("Lost an executor {} on slave {} with status {}", failure.getExecutorId(), failure.getAgentId(), failure.getStatus());
-    } else {
-      callWithLock(() -> slaveLost(failure.getAgentId()), "slaveLost");
-    }
-  }
-
-  @Override
-  public void error(String message) {
-    callWithLock(() -> {
-      LOG.error("Aborting due to error: {}", message);
-      abort.abort(AbortReason.MESOS_ERROR, Optional.absent());
-    }, "error");
-  }
-
-  @Override
-  public void heartbeat() {
-    LOG.debug("Heartbeat from mesos");
-  }
-
-  private FrameworkInfo buildFrameworkInfo() {
-    final FrameworkInfo.Builder frameworkInfoBuilder = FrameworkInfo.newBuilder()
-        .setCheckpoint(mesosConfiguration.isCheckpoint())
-        .setFailoverTimeout(mesosConfiguration.getFrameworkFailoverTimeout())
-        .setName(mesosConfiguration.getFrameworkName())
-        .setId(FrameworkID.newBuilder().setValue(mesosConfiguration.getFrameworkId()))
-        .setUser(mesosConfiguration.getFrameworkUser()); // https://issues.apache.org/jira/browse/MESOS-3747
-
-    if (configuration.getHostname().isPresent()) {
-      frameworkInfoBuilder.setHostname(configuration.getHostname().get());
-    }
-
-    // only set the web UI URL if it's fully qualified
-    if (singularityUriBase.startsWith("http://") || singularityUriBase.startsWith("https://")) {
-      if (configuration.getUiConfiguration().getRootUrlMode() == UIConfiguration.RootUrlMode.INDEX_CATCHALL) {
-        frameworkInfoBuilder.setWebuiUrl(singularityUriBase);
-      } else {
-        frameworkInfoBuilder.setWebuiUrl(singularityUriBase + UiResource.UI_RESOURCE_LOCATION);
-      }
-    }
-
-    if (mesosConfiguration.getFrameworkRole().isPresent()) {
-      frameworkInfoBuilder.setRole(mesosConfiguration.getFrameworkRole().get());
-    }
-
-    return frameworkInfoBuilder.build();
-  }
-
-  public void start() throws URISyntaxException {
-    subscribe(URI.create(configuration.getMesosConfiguration().getMasterUri()), buildFrameworkInfo());
-  }
-
-  private void callWithLock(Runnable function, String name) {
-    callWithLock(function, name, true);
-  }
-
-  private void callWithLock(Runnable function, String name, boolean ignoreIfNotRunning) {
-    if (!connected.get()) {
-      // This should never happen
-      LOG.warn("Attempted to process message, but not connected");
-    }
-    if (ignoreIfNotRunning && !isRunning()) {
-      LOG.info("Ignoring {} because scheduler isn't running ({})", name, state);
-      return;
-    }
-    final long start = lock.lock(name);
-    try {
-      function.run();
-    } catch (Throwable t) {
-      LOG.error("Scheduler threw an uncaught exception - exiting", t);
-      exceptionNotifier.notify(String.format("Scheduler threw an uncaught exception (%s)", t.getMessage()), t);
-      abort.abort(AbortReason.UNRECOVERABLE_ERROR, Optional.of(t));
-    } finally {
-      lock.unlock(name, start);
-    }
-  }
-
-  public void notifyStopping() {
-    LOG.info("Scheduler is moving to stopped, current state: {}", state);
-
-    state = SchedulerState.STOPPED;
-
-    leaderCacheCoordinator.stopLeaderCache();
-    close();
-
-    LOG.info("Scheduler now in state: {}", state);
-  }
-
-  public boolean isRunning() {
-    return state == SchedulerState.SUBSCRIBED;
-  }
-
-  public void setSubscribed() {
-    stateLock.lock();
-    try {
-      state = SchedulerState.SUBSCRIBED;
-    } finally {
-      stateLock.unlock();
-    }
-  }
-
-  public Optional<MasterInfo> getMaster() {
-    return Optional.fromNullable(masterInfo.get());
-  }
-
-  public void slaveLost(Protos.AgentID slaveId) {
-    LOG.warn("Lost a slave {}", slaveId);
-
-    slaveAndRackManager.slaveLost(slaveId);
-  }
-
-  public boolean isConnected() {
-    return connected.get();
-  }
-
-  public Optional<Long> getLastOfferTimestamp() {
-    return lastOfferTimestamp;
-  }
+  public abstract void killAndRecord(SingularityTaskId taskId, Optional<RequestCleanupType> requestCleanupType, Optional<TaskCleanupType> taskCleanupType, Optional<Long> originalTimestamp, Optional<Integer> retries, Optional<String> user);
 
   public void killAndRecord(SingularityTaskId taskId, RequestCleanupType requestCleanupType, Optional<String> user) {
     killAndRecord(taskId, Optional.of(requestCleanupType), Optional.<TaskCleanupType> absent(), Optional.<Long> absent(), Optional.<Integer> absent(), user);
@@ -432,47 +154,8 @@ public class SingularityMesosScheduler extends SingularityMesosSchedulerClient {
     killAndRecord(taskId, Optional.<RequestCleanupType> absent(), Optional.of(taskCleanupType), Optional.<Long> absent(), Optional.<Integer> absent(), user);
   }
 
-  public void killAndRecord(SingularityTaskId taskId, Optional<RequestCleanupType> requestCleanupType, Optional<TaskCleanupType> taskCleanupType, Optional<Long> originalTimestamp, Optional<Integer> retries, Optional<String> user) {
-    Preconditions.checkState(isRunning());
+  public abstract Optional<Long> getLastOfferTimestamp();
 
-    Optional<TaskCleanupType> maybeCleanupFromRequestAndTask = getTaskCleanupType(requestCleanupType, taskCleanupType);
-
-    if (maybeCleanupFromRequestAndTask.isPresent() && (maybeCleanupFromRequestAndTask.get() == TaskCleanupType.USER_REQUESTED_DESTROY || maybeCleanupFromRequestAndTask.get() == TaskCleanupType.REQUEST_DELETING)) {
-      Optional<SingularityTask> task = taskManager.getTask(taskId);
-      if (task.isPresent()) {
-        if (task.get().getTaskRequest().getDeploy().getCustomExecutorCmd().isPresent()) {
-          byte[] messageBytes = transcoder.toBytes(new SingularityTaskDestroyFrameworkMessage(taskId, user));
-          message(Message.newBuilder()
-              .setAgentId(task.get().getMesosTask().getAgentId())
-              .setExecutorId(task.get().getMesosTask().getExecutor().getExecutorId())
-              .setData(ByteString.copyFrom(messageBytes))
-              .build());
-        } else {
-          LOG.warn("Not using custom executor, will not send framework message to destroy task");
-        }
-      } else {
-        String message = String.format("No task data available to build kill task framework message for task %s", taskId);
-        exceptionNotifier.notify(message);
-        LOG.error(message);
-      }
-    }
-    kill(TaskID.newBuilder().setValue(taskId.toString()).build());
-
-    taskManager.saveKilledRecord(new SingularityKilledTaskIdRecord(taskId, System.currentTimeMillis(), originalTimestamp.or(System.currentTimeMillis()), requestCleanupType, taskCleanupType, retries.or(-1) + 1));
-  }
-
-  private Optional<TaskCleanupType> getTaskCleanupType(Optional<RequestCleanupType> requestCleanupType, Optional<TaskCleanupType> taskCleanupType) {
-    if (taskCleanupType.isPresent()) {
-      return taskCleanupType;
-    } else {
-      if (requestCleanupType.isPresent()) {
-        return requestCleanupType.get().getTaskCleanupType();
-      }
-      return Optional.absent();
-    }
-  }
-
-  public SchedulerState getState() {
-    return state;
-  }
+  public abstract void setSubscribed();
 }
+
