+++ SingularityService/src/test/java/com/hubspot/singularity/scheduler/SingularityTestModule.java
@@ -10,7 +10,6 @@ import javax.servlet.http.HttpServletRequest;
 
 import org.apache.curator.test.TestingServer;
 import org.eclipse.jetty.util.component.LifeCycle;
-import org.mockito.Mockito;
 import org.slf4j.LoggerFactory;
 
 import com.codahale.metrics.MetricRegistry;
@@ -54,8 +53,7 @@ import com.hubspot.singularity.hooks.LoadBalancerClient;
 import com.hubspot.singularity.mesos.OfferCache;
 import com.hubspot.singularity.mesos.SingularityMesosExecutorInfoSupport;
 import com.hubspot.singularity.mesos.SingularityMesosModule;
-import com.hubspot.singularity.mesos.SingularityMesosScheduler;
-import com.hubspot.singularity.mesos.SingularityMesosScheduler.SchedulerState;
+import com.hubspot.singularity.mesos.SingularityMesosSchedulerClient;
 import com.hubspot.singularity.mesos.SingularityOfferCache;
 import com.hubspot.singularity.resources.DeployResource;
 import com.hubspot.singularity.resources.PriorityResource;
@@ -180,24 +178,9 @@ public class SingularityTestModule implements Module {
             SingularityMesosExecutorInfoSupport logSupport = mock(SingularityMesosExecutorInfoSupport.class);
             binder.bind(SingularityMesosExecutorInfoSupport.class).toInstance(logSupport);
 
-            SingularityMesosScheduler mock = mock(SingularityMesosScheduler.class, Mockito.CALLS_REAL_METHODS);
-
-            SingularityMesosScheduler spy = Mockito.spy(mock);
-            try {
-              Mockito.doNothing().when(spy).start();
-            } catch (Exception e) {
-              // Skip
-            }
-            Mockito.doNothing().when(spy).decline(any());
-            Mockito.doNothing().when(spy).reconcile(any());
-            Mockito.doNothing().when(spy).kill(any());
-            Mockito.doNothing().when(spy).setSubscribed();
-            Mockito.doNothing().when(spy).accept(any(), anyList());
-            Mockito.doNothing().when(spy).frameworkMessage(any(), any(), any());
-            when(spy.isRunning()).thenReturn(true);
-            when(spy.getState()).thenReturn(SchedulerState.SUBSCRIBED);
-            when(spy.getLastOfferTimestamp()).thenReturn(Optional.absent());
-            binder.bind(SingularityMesosScheduler.class).toInstance(spy);
+            SingularityMesosSchedulerClient mockClient = mock(SingularityMesosSchedulerClient.class);
+            when(mockClient.isRunning()).thenReturn(true);
+            binder.bind(SingularityMesosSchedulerClient.class).toInstance(mockClient);
           }
         }));
